Update utilities function structure
